Move the marginnote counter to the context module
