+++ website/contexts/margin_note.js
@@ -1,4 +1,16 @@
 import React from "react";
 
+class MarginNoteCounter {
+  constructor() {
+    this.count = 1;
+  }
+
+  getAndUpdate() {
+    var oldCount = this.count;
+    this.count++;
+    return oldCount;
+  }
+}
+
 const MarginNoteContext = React.createContext();
-export default MarginNoteContext;
+export { MarginNoteContext, MarginNoteCounter };
