+++ website/pages/chapters/linear_regression.js
@@ -1,7 +1,7 @@
 import { BM, IM } from "../../components/latex";
-import { MarginNote, MarginNoteCounter } from "../../components/marginnote";
 
 import Chapter from "../../components/chapter";
+import MarginNote from "../../components/marginnote";
 import Video from "../../components/video";
 
 export default function LinearRegression() {
